+++ src/test/resources/features/Local/04_UpdateRecords.feature
@@ -2,13 +2,13 @@
 
 Feature: Trader Goods Profile - View or Update record journey
 
-  Scenario:As a returning user, With no historic records
-    Given I am on the 'Home' page with EORI GB123456789003
+  Scenario:As a new user, with no historic records, I want to view my goods profile
+    Given I am on the 'Home' page with EORI GB123456783000
     When I click on the View or update your goods records link
     Then I should be on the 'Goods profile' page
 
-  Scenario: As a new user, with historic data, View and update records
-    Given I am on the 'Home' page with EORI GB123456789099
+  Scenario: As a returning user, with historic data, I want to view and update records
+    Given I login to TGP Homepage with EORI number GB990000000008
     When I click on the View or update your goods records link
     Then I should be on the 'Previous movement records' page
     And I select continue
