+++ src/test/scala/uk/gov/hmrc/test/ui/pages/RequestAdvice/AdviceStartPage.scala
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.cucumber.stepdefs.Accreditation
+package uk.gov.hmrc.test.ui.pages.RequestAdvice
 
-import uk.gov.hmrc.test.ui.cucumber.stepdefs.BaseStepDef
-import uk.gov.hmrc.test.ui.pages.Accreditation.AdviceStartPage
+import uk.gov.hmrc.test.ui.pages.Base.Page
 
-class AdviceStartStepDef extends BaseStepDef {
+object AdviceStartPage extends Page {
 
-  Then("""^(?:I )?should be on the 'Asking HMRC for advice' page$""") { () =>
-    AdviceStartPage
-      .assertPage()
-  }
+  override def title(args: String*): String = "Asking HMRC for advice"
+  override def h1(args: String*): String    = "Asking HMRC for advice"
+  override def url(args: String*): String   = "/update-record/" + recordId + "/create-advice/start"
 }
