+++ src/test/scala/uk/gov/hmrc/test/ui/pages/RequestAdvice/AskEmailPage.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages.Accreditation
+package uk.gov.hmrc.test.ui.pages.RequestAdvice
 
 import uk.gov.hmrc.test.ui.pages.Base.Page
